% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passsword.generator.r
\name{passwords}
\alias{passwords}
\title{A password generator and safe}
\usage{
passwords(master, sites = c("work email", "online bank", "gmail"),
  changes = c(1, 1, 1, 1, 1, 1), how.many.letters = 9,
  how.many.LETTERS = 2, how.many.numbers = 2, how.many.special = 2)
}
\arguments{
\item{master}{your master password. Must be numerical with however many digits you want e.g. 7812362029}

\item{sites}{the names of things for which the password will be generated and saved}

\item{changes}{a binary vector to describe if you want change a password or not (same length as sites)}

\item{how.many.letters}{number of lower case letters in your passwords}

\item{how.many.LETTERS}{number of upper case letters in your passwords}

\item{how.many.numbers}{number of numbers in your passwords}

\item{how.many.special}{number of special characters in your passwords}
}
\description{
A password generator and safe
}
\examples{
 master.password=29852758
 passwords(master.password,sites=c("Work","Bank","Investment","gmail.com","yahoo.com","hotmail.com","visa","spotify"),changes=c(1,1,1,1,1,1,2,1))
}
\keyword{generator}
\keyword{password}
